A separate component for spinner v3
