+++ src/components/EventSummary.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react'
 import { Header, Segment, Grid, Embed, Image, Icon } from 'semantic-ui-react'
-import Custringloader  from './custringload'
+import Custringloader from '../containers/custringload'
 import ReactHtmlParser from 'react-html-parser'
 import moment from 'moment-timezone'
