+++ src/containers/GettingStartedPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Custringloader  from './custringload'
+import Custringloader from './custringload'
 import { Container } from 'semantic-ui-react'
 import axios from 'axios'
 import ReactHtmlParser from 'react-html-parser'
@@ -26,9 +26,9 @@ export default class GettingStartedPage extends Component {
         )
       } else {
         return (
-          
-            <Custringloader sizeUnit={'px'} size={200} color={'#34495E'} />
-          
+
+          <Custringloader sizeUnit={'px'} size={200} color={'#34495E'} />
+
         )
       }
     }
