+++ src/containers/PremiumMembershipPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import Custringloader  from './custringload'
+import Custringloader from './custringload'
 import { Container } from 'semantic-ui-react'
 import axios from 'axios'
 import ReactHtmlParser from 'react-html-parser'
@@ -24,8 +24,8 @@ export default class PremiumMembershipPage extends Component {
         </Container>
       )
     } else {
-      return (        
-          <Custringloader sizeUnit={'px'} size={200} color={'#34495E'} />       
+      return (
+        <Custringloader sizeUnit={'px'} size={200} color={'#34495E'} />
       )
     }
   }
