+++ src/containers/Subscriptions.js
@@ -7,7 +7,7 @@ import createBillingAgreement from '../actions/createBillingAgreement'
 import queryString from 'query-string'
 import membership from '../helpers/membershipInfo'
 import LoadingOverlay from 'react-loading-overlay'
-import Custringloader  from './custringload'
+import Custringloader from './custringload'
 import ErrorBoundary from '../components/ErrorBoundary'
 import StripeCheckout from 'react-stripe-checkout'
 import createStripeSubscription from '../actions/createStripeSubscription'
